[feat] javadoc 추가 및 필요없는 주석 제거 #39
