+++ src/App.tsx
@@ -9,8 +9,8 @@ import Layout from './modules/layout/Layout';
 import ModifySurvey from './modules/survey/modify/routers/ModifySurvey';
 import CreateationSurvey from './modules/survey/creation/routers/CreateationSurvey';
 import LoginDisplay from './modules/login/LoginDisplay';
-import MypageWriteSurvey from './modules/survey/mypage/MypageWriteSurvey';
-import MypageAttendSurvey from './modules/survey/mypage/MypageAttendSurvey';
+import MypageWriteSurvey from './modules/survey/mypage/components/MypageWriteSurvey';
+import MypageAttendSurvey from './modules/survey/mypage/components/MypageAttendSurvey';
 // import StatisticsPage from './modules/survey/statistic/StatisticsPage';
 import Main from './modules/survey/main/Main';
 import AttendSurvey from './modules/survey/attend/routers/AttendSurvey';
