--- src/modules/survey/attend/util/Animation.css
@@ -1,16 +1,0 @@
-/* 추가할 CSS */
-.hidden {
-  opacity: 0;
-  visibility: hidden;
-  transition:
-    opacity 0.5s ease-out,
-    visibility 0.5s ease-out;
-}
-
-.visible {
-  opacity: 1;
-  visibility: visible;
-  transition:
-    opacity 0.5s ease-in,
-    visibility 0.5s ease-in;
-}
