--- src/modules/survey/mypage/Mypage.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import Container from '@mui/material/Container';
-import { useNavigate } from 'react-router-dom';
-import { Button } from '@mui/material';
-
-function Mypage() {
-  const navigate = useNavigate();
-  const goBack = () => {
-    navigate('/survey/main');
-  };
-  return (
-    <Container maxWidth="md">
-      <h1>This is mypage</h1>
-      <Button onClick={goBack}>돌아가기</Button>
-    </Container>
-  );
-}
-
-export default Mypage;
