+++ src/modules/survey/mypage/components/MypageAttendSurvey.tsx
@@ -1,3 +1,10 @@
+/**
+ * 사용자의 마이페이지를 위한 컴포넌트입니다. 사용자가 참여한 설문 목록을 보여주고,
+ * 각 설문에 대해 상태를 필터링하거나 검색할 수 있는 기능을 제공합니다.
+ *
+ * @component
+ * @author 박창우
+ */
 import React, { useState, useEffect } from 'react';
 import Container from '@mui/material/Container';
 import axios from 'axios';
@@ -20,8 +27,6 @@ import Select, { SelectChangeEvent } from '@mui/material/Select';
 import Paper from '@mui/material/Paper';
 import InputBase from '@mui/material/InputBase';
 import Divider from '@mui/material/Divider';
-import { useNavigate } from 'react-router-dom';
-import Swal from 'sweetalert2';
 
 interface CardData {
   surveyNo: number;
@@ -34,11 +39,15 @@ interface CardData {
   surveyClosingAt: string;
   surveyPostAt: string;
   openStatusNo: number;
-  writer: string;
+  userNickname: string;
 }
 
-console.log('윈도우 온로드 체크');
-
+/**
+ * 설문의 상태 번호에 따라 해당하는 상태 텍스트를 반환합니다.
+ *
+ * @param {number} surveyStatusNo 설문의 상태 번호
+ * @returns {string} 설문 상태를 나타내는 텍스트
+ */
 function getStatusText(surveyStatusNo: number) {
   switch (surveyStatusNo) {
     case 1:
@@ -52,6 +61,12 @@ function getStatusText(surveyStatusNo: number) {
   }
 }
 
+/**
+ * 설문의 상태 번호에 따라 해당하는 칩 색상을 반환합니다.
+ *
+ * @param {number} surveyStatusNo 설문의 상태 번호
+ * @returns {string} Material UI 칩 컴포넌트에 사용될 색상
+ */
 function getChipColor(surveyStatusNo: number) {
   switch (surveyStatusNo) {
     case 1:
@@ -63,6 +78,12 @@ function getChipColor(surveyStatusNo: number) {
   }
 }
 
+/**
+ * 설문의 상태 번호에 따라 해당하는 카드 색상을 반환합니다.
+ *
+ * @param {number} surveyStatusNo 설문의 상태 번호
+ * @returns {string} 카드 배경에 사용될 색상
+ */
 function getCardColor(surveyStatusNo: number) {
   switch (surveyStatusNo) {
     case 1:
@@ -83,72 +104,20 @@ function Mypage() {
 
   const [state, setState] = useState('전체');
 
-  const [isUpdateData, setIsUpdateData] = useState<boolean>(false);
-
   const userNo = 1;
 
   const [searchQuery, setSearchQuery] = useState('');
-
-  const naviagte = useNavigate();
-
-  /**
-   * 설문수정 페이지로 리다이렉트 시키는 onClick 메서드 입니다.
-   *
-   * @param surveyNo 설문 번호
-   * @author 강명관
-   */
-  const handleClickSurveyModify = (surveyNo: number) => {
-    naviagte(`/survey/modify/${surveyNo}`);
-  };
-
-  /**
-   * 설문을 진행 상태에서 게시 상태로 변경하는 API Call 하는 메서드 입니다.
-   *
-   * @param surveyNo 설문 번호
-   * @author 강명관
-   */
-  const handleClickPostSurvey = async (surveyNo: number) => {
-    try {
-      const response = await axios.put(
-        `http://localhost:8080/api/surveys/${surveyNo}/post`
-      );
-
-      if (response.status === 200) {
-        setOpenModal(false);
-        setIsUpdateData(true);
-
-        Swal.fire({
-          icon: 'success',
-          title: '설문 게시가 완료되었습니다!',
-        });
-      } else {
-        setOpenModal(false);
-
-        Swal.fire({
-          icon: 'error',
-          title: '설문 게시에 실패했습니다.',
-          text: `설문의 상태 혹은 설문의 마감일을 확인해주세요.`,
-        });
-      }
-    } catch (error) {
-      console.error(error);
-
-      setOpenModal(false);
-      Swal.fire({
-        icon: 'error',
-        title: '설문 게시에 실패했습니다.',
-        text: `설문의 상태 혹은 설문의 마감일을 확인해주세요.`,
-      });
-    }
-  };
 
   const handleChange = (event: SelectChangeEvent) => {
     setState(event.target.value);
   };
 
+  /**
+   * 서버로부터 사용자가 참여한 설문 데이터를 가져와 상태를 업데이트합니다.
+   */
   const fetchCardData = () => {
     axios
-      .get(`http://localhost:8080/api/my-surveys/${userNo}/write-surveys`)
+      .get(`http://localhost:8080/api/my-surveys/${userNo}/attend-surveys`)
       .then((response) => {
         const cardData: CardData[] = response.data.content;
 
@@ -190,15 +159,14 @@ function Mypage() {
   };
 
   useEffect(() => {
-    if (isUpdateData) {
-      fetchCardData();
-    }
-  }, [isUpdateData]);
-
-  useEffect(() => {
     fetchCardData();
   }, [userNo, state, searchQuery]);
 
+  /**
+   * 사용자가 검색 입력란에 입력할 때마다 호출되어 검색 쿼리 상태를 업데이트합니다.
+   *
+   * @param {React.ChangeEvent<HTMLInputElement>} event 입력 이벤트
+   */
   const handleSearchChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     const newSearchQuery = event.target.value;
     setSearchQuery(newSearchQuery);
@@ -206,16 +174,27 @@ function Mypage() {
     fetchCardData();
   };
 
+  /**
+   * 카드를 클릭했을 때 모달을 열어 해당 설문의 상세 정보를 보여줍니다.
+   *
+   * @param {CardData} card 사용자가 클릭한 설문 데이터
+   */
   const openCardModal = (card: CardData) => {
     setSelectedCard(card);
     setOpenModal(true);
   };
 
+  /**
+   * 모달을 닫고 선택된 설문의 상태를 초기화합니다.
+   */
   const closeCardModal = () => {
     setSelectedCard(null);
     setOpenModal(false);
   };
 
+  /**
+   * 사용자가 설문 삭제 버튼을 클릭했을 때 호출되어, 설문을 삭제합니다.
+   */
   const handleDeleteClick = () => {
     if (selectedCard) {
       if (window.confirm('작성 중인 설문을 삭제하시겠습니까?')) {
@@ -251,7 +230,7 @@ function Mypage() {
           fontSize: '25px',
         }}
       >
-        내가 작성한 설문 목록
+        내가 참여한 설문 목록
       </h1>
 
       <Box
@@ -452,6 +431,15 @@ function Mypage() {
                     </React.Fragment>
                   ))}
                 </Typography>
+                <Typography
+                  sx={{
+                    fontSize: 12,
+                    fontWeight: 600,
+                    paddingTop: '12px',
+                  }}
+                >
+                  작성자: {card.userNickname}
+                </Typography>
               </CardContent>
             </Card>
           </div>
@@ -495,6 +483,7 @@ function Mypage() {
                 : ''}
             </p>
 
+            <p>작성자: {selectedCard ? selectedCard.userNickname : ''}</p>
             <p>태그: {selectedCard ? selectedCard.tagNames.join(', ') : ''}</p>
             <p>참석자 수: {selectedCard ? selectedCard.attendeeCount : ''}</p>
             <p id="modal-description">
@@ -503,17 +492,9 @@ function Mypage() {
 
             {selectedCard && selectedCard.surveyStatusNo === 1 && (
               <>
-                <Button
-                  onClick={() => handleClickSurveyModify(selectedCard.surveyNo)}
-                >
-                  수정하기
-                </Button>
+                <Button>수정하기</Button>
                 <Button onClick={handleDeleteClick}>삭제하기</Button>
-                <Button
-                  onClick={() => handleClickPostSurvey(selectedCard.surveyNo)}
-                >
-                  게시하기
-                </Button>
+                <Button>게시하기</Button>
               </>
             )}
