+++ src/modules/survey/mypage/components/MypageWriteSurvey.tsx
@@ -1,3 +1,10 @@
+/**
+ * 사용자의 마이페이지 컴포넌트입니다. 사용자가 작성하거나 참여한 설문의 목록을 보여주고,
+ * 상태에 따라 필터링하거나 검색할 수 있는 기능을 제공합니다.
+ *
+ * @returns {React.ReactElement} 마이페이지 컴포넌트를 렌더링하는 React 엘리먼트
+ * @author 박창우
+ */
 import React, { useState, useEffect } from 'react';
 import Container from '@mui/material/Container';
 import axios from 'axios';
@@ -20,6 +27,8 @@ import Select, { SelectChangeEvent } from '@mui/material/Select';
 import Paper from '@mui/material/Paper';
 import InputBase from '@mui/material/InputBase';
 import Divider from '@mui/material/Divider';
+import { useNavigate } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 interface CardData {
   surveyNo: number;
@@ -32,9 +41,17 @@ interface CardData {
   surveyClosingAt: string;
   surveyPostAt: string;
   openStatusNo: number;
-  userNickname: string;
+  writer: string;
 }
 
+console.log('윈도우 온로드 체크');
+
+/**
+ * 설문 상태 번호에 따라 상태 텍스트를 반환합니다.
+ *
+ * @param {number} surveyStatusNo - 설문의 상태 번호
+ * @returns {string} 해당 상태 번호에 대한 텍스트 설명
+ */
 function getStatusText(surveyStatusNo: number) {
   switch (surveyStatusNo) {
     case 1:
@@ -48,6 +65,12 @@ function getStatusText(surveyStatusNo: number) {
   }
 }
 
+/**
+ * 설문 상태 번호에 따라 Material UI Chip 컴포넌트의 색상을 반환합니다.
+ *
+ * @param {number} surveyStatusNo - 설문의 상태 번호
+ * @returns {string} 해당 상태에 맞는 색상 이름
+ */
 function getChipColor(surveyStatusNo: number) {
   switch (surveyStatusNo) {
     case 1:
@@ -59,6 +82,12 @@ function getChipColor(surveyStatusNo: number) {
   }
 }
 
+/**
+ * 설문 상태 번호에 따라 카드 배경색을 반환합니다.
+ *
+ * @param {number} surveyStatusNo - 설문의 상태 번호
+ * @returns {string} 해당 상태에 맞는 배경색
+ */
 function getCardColor(surveyStatusNo: number) {
   switch (surveyStatusNo) {
     case 1:
@@ -79,17 +108,75 @@ function Mypage() {
 
   const [state, setState] = useState('전체');
 
+  const [isUpdateData, setIsUpdateData] = useState<boolean>(false);
+
   const userNo = 1;
 
   const [searchQuery, setSearchQuery] = useState('');
+
+  const naviagte = useNavigate();
+
+  /**
+   * 설문수정 페이지로 리다이렉트 시키는 onClick 메서드 입니다.
+   *
+   * @param surveyNo 설문 번호
+   * @author 강명관
+   */
+  const handleClickSurveyModify = (surveyNo: number) => {
+    naviagte(`/survey/modify/${surveyNo}`);
+  };
+
+  /**
+   * 설문을 진행 상태에서 게시 상태로 변경하는 API Call 하는 메서드 입니다.
+   *
+   * @param surveyNo 설문 번호
+   * @author 강명관
+   */
+  const handleClickPostSurvey = async (surveyNo: number) => {
+    try {
+      const response = await axios.put(
+        `http://localhost:8080/api/surveys/${surveyNo}/post`
+      );
+
+      if (response.status === 200) {
+        setOpenModal(false);
+        setIsUpdateData(true);
+
+        Swal.fire({
+          icon: 'success',
+          title: '설문 게시가 완료되었습니다!',
+        });
+      } else {
+        setOpenModal(false);
+
+        Swal.fire({
+          icon: 'error',
+          title: '설문 게시에 실패했습니다.',
+          text: `설문의 상태 혹은 설문의 마감일을 확인해주세요.`,
+        });
+      }
+    } catch (error) {
+      console.error(error);
+
+      setOpenModal(false);
+      Swal.fire({
+        icon: 'error',
+        title: '설문 게시에 실패했습니다.',
+        text: `설문의 상태 혹은 설문의 마감일을 확인해주세요.`,
+      });
+    }
+  };
 
   const handleChange = (event: SelectChangeEvent) => {
     setState(event.target.value);
   };
 
+  /**
+   * 선택된 상태나 검색 쿼리에 따라 필터링된 설문 데이터를 서버로부터 가져옵니다.
+   */
   const fetchCardData = () => {
     axios
-      .get(`http://localhost:8080/api/my-surveys/${userNo}/attend-surveys`)
+      .get(`http://localhost:8080/api/my-surveys/${userNo}/write-surveys`)
       .then((response) => {
         const cardData: CardData[] = response.data.content;
 
@@ -131,9 +218,20 @@ function Mypage() {
   };
 
   useEffect(() => {
+    if (isUpdateData) {
+      fetchCardData();
+    }
+  }, [isUpdateData]);
+
+  useEffect(() => {
     fetchCardData();
   }, [userNo, state, searchQuery]);
 
+  /**
+   * 검색 입력란의 값이 변경될 때 호출되며, 검색 쿼리 상태를 업데이트합니다.
+   *
+   * @param {React.ChangeEvent<HTMLInputElement>} event - 입력 변경 이벤트
+   */
   const handleSearchChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     const newSearchQuery = event.target.value;
     setSearchQuery(newSearchQuery);
@@ -141,16 +239,27 @@ function Mypage() {
     fetchCardData();
   };
 
+  /**
+   * 설문 카드를 클릭했을 때 모달을 열어 상세 정보를 보여줍니다.
+   *
+   * @param {CardData} card - 선택된 카드 데이터
+   */
   const openCardModal = (card: CardData) => {
     setSelectedCard(card);
     setOpenModal(true);
   };
 
+  /**
+   * 모달을 닫고 선택된 설문 상태를 초기화합니다.
+   */
   const closeCardModal = () => {
     setSelectedCard(null);
     setOpenModal(false);
   };
 
+  /**
+   * 삭제 버튼 클릭 시 작성 중인 설문을 삭제합니다.
+   */
   const handleDeleteClick = () => {
     if (selectedCard) {
       if (window.confirm('작성 중인 설문을 삭제하시겠습니까?')) {
@@ -186,7 +295,7 @@ function Mypage() {
           fontSize: '25px',
         }}
       >
-        내가 참여한 설문 목록
+        내가 작성한 설문 목록
       </h1>
 
       <Box
@@ -387,15 +496,6 @@ function Mypage() {
                     </React.Fragment>
                   ))}
                 </Typography>
-                <Typography
-                  sx={{
-                    fontSize: 12,
-                    fontWeight: 600,
-                    paddingTop: '12px',
-                  }}
-                >
-                  작성자: {card.userNickname}
-                </Typography>
               </CardContent>
             </Card>
           </div>
@@ -439,7 +539,6 @@ function Mypage() {
                 : ''}
             </p>
 
-            <p>작성자: {selectedCard ? selectedCard.userNickname : ''}</p>
             <p>태그: {selectedCard ? selectedCard.tagNames.join(', ') : ''}</p>
             <p>참석자 수: {selectedCard ? selectedCard.attendeeCount : ''}</p>
             <p id="modal-description">
@@ -448,9 +547,17 @@ function Mypage() {
 
             {selectedCard && selectedCard.surveyStatusNo === 1 && (
               <>
-                <Button>수정하기</Button>
+                <Button
+                  onClick={() => handleClickSurveyModify(selectedCard.surveyNo)}
+                >
+                  수정하기
+                </Button>
                 <Button onClick={handleDeleteClick}>삭제하기</Button>
-                <Button>게시하기</Button>
+                <Button
+                  onClick={() => handleClickPostSurvey(selectedCard.surveyNo)}
+                >
+                  게시하기
+                </Button>
               </>
             )}
